Delete type convertion

+++ get_hh_salaries.py
@@ -43,7 +43,7 @@ def get_hh_summary_vacancies(vacancies_pages):
             if salary:
                 salaries.append(salary)
     try:
-        average_salary = int((sum(salaries) / len(salaries)))
+        average_salary = sum(salaries) / len(salaries)
     except ZeroDivisionError:
         average_salary = 0
     vacancies_processed = len(salaries)
